backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
colorlog>=6.7.0
tqdm>=4.65.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
numpy>=1.24.0
imagehash>=4.3.1
requests>=2.31.0 